rename views/forms to views/setup
