+++ corehq/apps/data_cleaning/urls.py
@@ -8,7 +8,7 @@ from corehq.apps.data_cleaning.views.tables import (
     CleanCasesTableView,
     CaseCleaningTasksTableView,
 )
-from corehq.apps.data_cleaning.views.forms import (
+from corehq.apps.data_cleaning.views.setup import (
     SetupCaseSessionFormView,
 )
